[20424] First production beta!
[20424] is the first version of the app which could be considered a
working “beta.” It has all of the functionality of the first working
version, but it’s lacking extensive testing and a few touch-ups here
and there.

You may have noticed that there was no [20423]. Well, there was
actually. It took me 4 hours of straight work to write it. And then I
lost it all. So I took another 4 hours to re-write it. And now we’re
here. [20423] is now an unlucky build number, so we’re skipping over it.